backend/requirements.txt
Django==4.2.9
psycopg[binary]
tzdata
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter
whitenoise
django-ckeditor-5
django-cors-headers